requirements.txt
requests==2.32.3
feedparser==6.0.11
beautifulsoup4==4.12.3
lxml==5.3.0
Pillow==10.4.0
pillow-avif-plugin==1.4.5
tenacity==9.0.0
pytest==8.3.3
python-dotenv==1.0.0
newspaper3k==0.2.8
trafilatura==1.8.0
readability-lxml==0.8.1